Add style guide for the Title and EditTitle components
