+++ src/components/EditTitle/EditTitle.md
@@ -1,0 +1,21 @@
+**Default**
+
+```js
+<EditTitle
+  title='The Title'
+  level={1}
+  handleClick={() => alert('Clicked')}
+  editable
+/>
+```
+
+**Not Editable**
+
+```js
+<EditTitle
+  title='The Title'
+  level={1}
+  handleClick={() => alert('Clicked')}
+  editable={false}
+/>
+```
